added checksum comparison to file sync service

+++ FolderSyncTool.App/FileSync/Service/FileSyncService.cs
@@ -1,19 +1,15 @@
-﻿using System;
-using System.IO;
+﻿using System.Security.Cryptography;
 
 namespace FolderSyncTool.App.FileSync.Service
 {
     public class FileSyncService : IFileSyncService
     {
-        private const string SourceFolderPath = "E:\\SyncTest\\Source";
-        private const string ReplicaFolderPath = "E:\\SyncTest\\Replica";
-
-        public void Sync()
+        public void Sync(string sourcePath, string targetPath)
         {
-            CopyDirectory(SourceFolderPath, ReplicaFolderPath);
+            ReplicateDirectory(sourcePath, targetPath);
         }
 
-        public static void CopyDirectory(string sourcePath, string targetPath, bool recursive = true)
+        private static void ReplicateDirectory(string sourcePath, string targetPath)
         {
             if(string.IsNullOrEmpty(sourcePath) || string.IsNullOrEmpty(targetPath))
             {
@@ -27,18 +23,54 @@ namespace FolderSyncTool.App.FileSync.Service
 
             Directory.CreateDirectory(targetPath);
 
-            foreach(var filePath in Directory.GetFiles(sourcePath))
-            {
-                string targetFilePath = Path.Combine(targetPath, Path.GetFileName(filePath));
-                File.Copy(filePath, targetFilePath, true);
-            }
-
-            if (recursive == false) return;
+            ClaerTargetFiles(sourcePath, targetPath);
+            ReplicateSourceFiles(sourcePath, targetPath);
 
             foreach(var sourceSubPath in Directory.GetDirectories(sourcePath))
             {
                 string targetSubPath = Path.Combine(targetPath, Path.GetFileName(sourceSubPath));
-                CopyDirectory(sourceSubPath, targetSubPath);
+                ReplicateDirectory(sourceSubPath, targetSubPath);
+            }
+        }
+
+        private static void ClaerTargetFiles(string sourcePath, string targetPath)
+        {
+            foreach (var filePath in Directory.GetFiles(targetPath))
+            {
+                string sourceFilePath = Path.Combine(sourcePath, Path.GetFileName(filePath));
+
+                if (File.Exists(sourceFilePath)) continue;
+
+                File.Delete(filePath);
+            }
+        }
+
+        private static void ReplicateSourceFiles(string sourcePath, string targetPath)
+        {
+            foreach (var filePath in Directory.GetFiles(sourcePath))
+            {
+                string targetFilePath = Path.Combine(targetPath, Path.GetFileName(filePath));
+
+                if (File.Exists(targetFilePath) && !IsFileChanged(filePath, targetFilePath)) continue;
+
+                File.Copy(filePath, targetFilePath, true);
+            }
+        }
+
+        private static bool IsFileChanged(string sourceFile, string targetFile)
+        {
+            return !string.Equals(CheckMD5(sourceFile), CheckMD5(targetFile));
+        }
+
+        private static string CheckMD5(string filePath)
+        {
+            using(var md5 = MD5.Create())
+            {
+                using (var stream = File.OpenRead(filePath))
+                {
+                    var hash = md5.ComputeHash(stream);
+                    return BitConverter.ToString(hash).Replace("-", "").ToLower();
+                }
             }
         }
     }
